refactor and test sgat
