+++ deeprank_gnn/sGAT.py
@@ -16,7 +16,48 @@ from torch_geometric.nn import max_pool_x
 from .community_pooling import get_preloaded_cluster, community_pooling
 
 
-class sGraphAttention(torch.nn.Module):
+class sGAT(torch.nn.Module):
+
+    def __init__(self, input_shape, output_shape=1, input_shape_edge=None):
+        super(sGAT, self).__init__()
+
+        self.conv1 = sGraphAttentionLayer(input_shape, 16)
+        self.conv2 = sGraphAttentionLayer(16, 32)
+
+        self.fc1 = torch.nn.Linear(32, 64)
+        self.fc2 = torch.nn.Linear(64, output_shape)
+
+        self.clustering = 'mcl'
+
+    def forward(self, data):
+
+        act = nn.Tanhshrink()
+        act = F.relu
+        #act = nn.LeakyReLU(0.25)
+
+        # first conv block
+        data.x = act(self.conv1(
+            data.x, data.edge_index, data.edge_attr))
+        cluster = get_preloaded_cluster(data.cluster0, data.batch)
+        data = community_pooling(cluster, data)
+
+        # second conv block
+        data.x = act(self.conv2(
+            data.x, data.edge_index, data.edge_attr))
+        cluster = get_preloaded_cluster(data.cluster1, data.batch)
+        x, batch = max_pool_x(cluster, data.x, data.batch)
+
+        # FC
+        x = scatter_mean(x, batch, dim=0)
+        x = act(self.fc1(x))
+        x = self.fc2(x)
+        #x = F.dropout(x, training=self.training)
+
+        return x
+        # return F.relu(x)
+
+
+class sGraphAttentionLayer(torch.nn.Module):
 
     """
     This is a new layer that is similar to the graph attention network but simpler
@@ -37,7 +78,7 @@ class sGraphAttention(torch.nn.Module):
                  out_channels,
                  bias=True):
 
-        super(sGraphAttention, self).__init__()
+        super(sGraphAttentionLayer, self).__init__()
 
         self.in_channels = in_channels
         self.out_channels = out_channels
@@ -97,42 +138,82 @@ class sGraphAttention(torch.nn.Module):
                                    self.out_channels)
 
 
-class sGAT(torch.nn.Module):
+class sGraphAttentionConvLayer(torch.nn.Module):
 
-    def __init__(self, input_shape, output_shape=1, input_shape_edge=None):
-        super(sGAT, self).__init__()
+    """
+    This is a new layer that is similar to the graph attention network but simpler
+    z_i =  1 / Ni \Sum_j a_ij * [x_i || x_j] * W + b_i
+    || is the concatenation operator: [1,2,3] || [4,5,6] = [1,2,3,4,5,6]
+    Ni is the number of neighbor of node i
+    \Sum_j runs over the neighbors of node i
+    a_ij is the edge attribute between node i and j
+    Args:
+        in_channels (int): Size of each input sample.
+        out_channels (int): Size of each output sample.
+        bias (bool, optional): If set to :obj:`False`, the layer will not learn
+            an additive bias. (default: :obj:`True`)
+    """
 
-        self.conv1 = sGraphAttention(input_shape, 16)
-        self.conv2 = sGraphAttention(16, 32)
+    def __init__(self,
+                 in_channels,
+                 out_channels,
+                 bias=True):
 
-        self.fc1 = torch.nn.Linear(32, 64)
-        self.fc2 = torch.nn.Linear(64, output_shape)
+        super(sGraphAttentionConvLayer, self).__init__()
 
-        self.clustering = 'mcl'
+        self.in_channels = in_channels
+        self.out_channels = out_channels
 
-    def forward(self, data):
+        self.weight = Parameter(
+            torch.Tensor(2 * in_channels, out_channels))
 
-        act = nn.Tanhshrink()
-        act = F.relu
-        #act = nn.LeakyReLU(0.25)
+        if bias:
+            self.bias = Parameter(torch.Tensor(out_channels))
+        else:
+            self.register_parameter('bias', None)
 
-        # first conv block
-        data.x = act(self.conv1(
-            data.x, data.edge_index, data.edge_attr))
-        cluster = get_preloaded_cluster(data.cluster0, data.batch)
-        data = community_pooling(cluster, data)
+        self.reset_parameters()
 
-        # second conv block
-        data.x = act(self.conv2(
-            data.x, data.edge_index, data.edge_attr))
-        cluster = get_preloaded_cluster(data.cluster1, data.batch)
-        x, batch = max_pool_x(cluster, data.x, data.batch)
+    def reset_parameters(self):
+        size = 2 * self.in_channels
+        uniform(size, self.weight)
+        uniform(size, self.bias)
 
-        # FC
-        x = scatter_mean(x, batch, dim=0)
-        x = act(self.fc1(x))
-        x = self.fc2(x)
-        #x = F.dropout(x, training=self.training)
+    def forward(self, x, edge_index, edge_attr):
 
-        return x
-        # return F.relu(x)
+        #print('weight : ', torch.sum(self.weight))
+
+        row, col = edge_index
+        num_node = len(x)
+        edge_attr = edge_attr.unsqueeze(
+            -1) if edge_attr.dim() == 1 else edge_attr
+
+        # create edge feature by concatenating node feature
+        alpha = torch.cat([x[row], x[col]], dim=-1)
+
+        # multiply the edge features with the fliter
+        alpha = torch.mm(alpha, self.weight)
+
+        # multiply each edge features with the corresponding dist
+        alpha = edge_attr*alpha
+
+        # scatter the resulting edge feature to get node features
+        out = torch.zeros(
+            num_node, self.out_channels).to(alpha.device)
+        out = scatter_mean(alpha, row, dim=0, out=out)
+
+        # if the graph is undirected and (i,j) and (j,i) are both in
+        # the edge_index then we do not need to have that second line
+        # or we count everythong twice
+        #out = scatter_mean(alpha,col,dim=0,out=out)
+
+        # add the bias
+        if self.bias is not None:
+            out = out + self.bias
+
+        return out
+
+    def __repr__(self):
+        return '{}({}, {})'.format(self.__class__.__name__,
+                                   self.in_channels,
+                                   self.out_channels)
